rename: getEvents -> getPopups로 수정
