+++ src/app/popup/search/page.tsx
@@ -1,5 +1,5 @@
 import WrapperPopupSearch from '@/features/popup-search/WrapperPopupSearch';
-import { getEvents, GetEventsParams } from '@/services/getEvents';
+import { GetEventsParams, getPopups } from '@/services/getPopups';
 
 interface PageProps {
   searchParams: Promise<{ zone?: string; tags?: string }>;
@@ -10,13 +10,13 @@ export default async function Page({ searchParams }: PageProps) {
   const selectedTags = tags ? [tags.split(',')[0]] : [];
 
   // 초기접속시 보여줄 이벤트 목록
-  const { events, totalCount } = await getEvents({
+  const { events, totalCount } = await getPopups({
     zone: zone ?? null,
     tags: selectedTags,
   } as GetEventsParams);
 
   // 전체 이벤트(전체 조회, 맵에 표시하기 위함)
-  const { events: fullEvents } = await getEvents({
+  const { events: fullEvents } = await getPopups({
     zone: zone ?? null,
     tags: selectedTags,
